backend/requirements.txt
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.7.4
cffi==1.17.0
channels==4.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
daphne==4.1.2
Django==5.1
django-channels==0.7.0
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
email_validator==2.2.0
Faker==28.0.0
fastapi==0.112.2
fonttools==4.55.0
greenlet==3.0.3
h11==0.14.0
hyperlink==21.0.0
idna==3.8
imageio==2.36.0
importlib_metadata==8.4.0
incremental==24.7.2
Jinja2==3.1.4
kiwisolver==1.4.7
Mako==1.3.8
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==2.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyee==11.1.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
pyppeteer==2.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytube==15.0.0
pyvista==0.44.2
requests==2.32.3
requests-oauthlib==2.0.0
scooby==0.10.0
selenium==4.23.1
selenium-authenticated-proxy==1.1.2
service-identity==24.2.0
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqladmin==0.18.0
SQLAlchemy==2.0.32
sqlparse==0.5.1
starlette==0.38.2
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
uvicorn==0.30.6
vtk==9.3.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.1.2
zipp==3.20.1
zope.interface==7.1.1
